Add files via upload
Consider N members are standing in a circle. The 1st person has a gun, shoots the 3rd person, hands the gun to the 5th member, who then shoots the 7th and then passes the gun to the 9th member. This cycle goes on until only one member remains. Find out which will be the last person standing for some N?